Set default timeout to 60 seconds
This only affects the first call. By default, a GenServer fails after 5
seconds, but since this is intended to cache slow commands, we need to
give it time to run them. I have it set to 1 minute for now so that I
don't have to worry about what happens if the command never returns (for
example `cat`), but `:infinity` may be a better choice, which would
allow it to wait indefinitely.

If I set it to infinity, it would probably be good to handle client
disconnect by killing the underlying process